requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.17.0
transformers==4.39.3
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
numpy==1.26.4
PyPDF2==3.0.1
pdf2image==1.17.0
pymupdf==1.23.26
pillow==10.2.0
azure-ai-formrecognizer==3.3.0
azure-core==1.30.1
requests==2.31.0
jinja2==3.1.3
httpx==0.27.0
pytest==8.1.1
